// src/app/Data/Interfaces/Iuser.ts
export interface Iuser {    
       userid: number;
       firstname: string;
       lastname: string;
       username: string;
       email: string;
       password: string;
       isPrimeUser: boolean;
       watchLater: string;
       favourites: string;
       watched: string;
       isAdmin: boolean ;
       dateOfBirth?: Date;   
    }
  